# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for Text en/decryption requests",
        "title": "DevIntApp",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/text/pic/{Id}": {
            "post": {
                "description": "Change text`s picture using it's ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Change picture By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text Id",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File",
                        "name": "img",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Picture was changed sucessfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/text_to_order/{Id}": {
            "post": {
                "description": "This endpoint allows you to add a text to a order by it's ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Add text to order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text Id",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Text added successfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.AddTextToOrderResponce"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/texts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all texts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Get all texts",
                "responses": {
                    "200": {
                        "description": "List of texts retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetAllTextsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Create text with properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Create text",
                "parameters": [
                    {
                        "description": "Text data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateTextRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateTextResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/texts/{Id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get info about text using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Get text by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text Id",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetTextResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update text using it's ID with parametres",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Update text by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text Id",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update text data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateTextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteTextResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete text using it's ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "texts"
                ],
                "summary": "Delete text by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text Id",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteTextResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ds.TextToEncOrDec": {
            "type": "object",
            "properties": {
                "byte_len": {
                    "type": "integer"
                },
                "enc": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "schemas.AddTextToOrderResponce": {
            "type": "object",
            "properties": {
                "messageResponse": {
                    "type": "string"
                },
                "orderId": {
                    "type": "integer"
                },
                "textId": {
                    "type": "integer"
                }
            }
        },
        "schemas.CreateTextRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "$ref": "#/definitions/ds.TextToEncOrDec"
                }
            }
        },
        "schemas.CreateTextResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                }
            }
        },
        "schemas.DeleteTextResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                }
            }
        },
...
        "schemas.GetAllTextsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "text_req_ID": {
                    "type": "integer"
                },
                "texts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.TextToEncOrDec"
                    }
                }
            }
        },
        "schemas.GetTextResponse": {
            "type": "object",
            "properties": {
                "text": {
                    "$ref": "#/definitions/ds.TextToEncOrDec"
                }
            }
        },
...
        "schemas.ResponseMessage": {
            "type": "object"
        },
        "schemas.UpdateTextRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "text": {
                    "$ref": "#/definitions/ds.TextToEncOrDec"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}